requirements.txt
dlib==19.24.2
Flask==3.0.3
imutils==0.5.4
opencv_python==4.10.0.84
openpyxl==3.1.5
Pillow==10.4.0
Requests==2.32.3
Werkzeug==3.0.3
numpy==1.26.4
redis==5.0.8
